Use pre-computed sizeBytes for catalog size and consolidate time formatting
Replace the broken calculateTotalStorage (which iterated all RemoteStorage
files) with a simple Dexie query summing article.sizeBytes, leveraging the
new per-article size tracking from ingestion.

Move formatReadTime() from PreferenceScreen.tsx to lib/src/lib.ts as a
shared utility. Update generateInfoForCard() to use it. Make format more
verbose: "5m" -> "5 min", "2h 30m" -> "2 hr 30 min", etc.
